requirements.txt
alpaca-py
pandas>=2.0.0
pandera>=0.18.0
numpy>=1.24.0
pyarrow>=14.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
PyYAML>=6.0
scikit-learn>=1.3.0
joblib>=1.3.0
pydantic-settings
pytz


# Database dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# UI dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
